Experimenting with IPC
