+++ index.html
@@ -8,7 +8,16 @@
 <link rel="stylesheet" href="css/style.css" type="text/css">
 <link rel="stylesheet" href="css/throbber.css" type="text/css">
 <link rel="stylesheet" href="css/breadcrumb.css" type="text/css">
+<script>
 
+var ipc = require('ipc')
+
+
+ ipc.send('ducomms', 'frommm index 123')
+ipc.on('ducomms', function(arg) {
+  console.log('index Renderer argum', arg)
+})
+</script>
 <div id="loading">
   <div class="throbber-loader">
     Loading…
@@ -29,7 +38,7 @@
 </div>
 
 <script src="js/utils.js"></script>
-<script src="js/du.js"></script>
+<!-- <script src="js/du.js"></script> -->
 <script src="js/bipartition_sunburst.js"></script>
 <!-- <script src="js/treemap.js"></script> -->
 </body>
