+++ js/du.js
@@ -4,7 +4,6 @@
 const fs = require('fs')
 const path = require('path')
 
-let target = '../../..'
 let counter = 0;
 
 /* Synchronous File System descender */
@@ -125,86 +124,6 @@ function descendFS(options, done) {
 	})
 }
 
-var INCREMENTAL_INTERVAL = 5000
-
-console.log('lets go');
-console.time('async2')
-// loading.style.display = 'inline-block'
-
-function complete() {
-    console.log("Scan completed", counter, "files");
-    console.timeEnd('async2')
-    clearTimeout(checker)
-    // loading.style.display = 'none'
-
-    console.log(json);
-
-    console.time('write')
-    fs.writeFileSync('test.json', JSON.stringify(json))
-    console.timeEnd('write')
-
-    onJson(null, clone2(json))
-};
-
-let checker
-
-var json = {};
-
-function updatePartialFS() {
-	clearTimeout(checker)
-	console.log('scanning...');
-	console.time('clone')
-	let cloneJson = clone2(json)
-	console.log(cloneJson)
-	console.timeEnd('clone')
-	onJson(null, cloneJson)
-	checker = setTimeout(updatePartialFS, INCREMENTAL_INTERVAL)
-}
-
-target = path.resolve(target)
-console.log('Scanning target', target)
-
-// d3.json("flare.json", onJson);
-
-setTimeout( () => {
-	// jsonFS(target, null, json)
-	// complete()
-
-	descendFS({
-		parent: target,
-		node: json,
-		onprogress: window.onProcess
-	}, complete)
-
-	// updatePartialFS();
-
-	setTimeout(updatePartialFS, INCREMENTAL_INTERVAL);
-
-	// for testing purposes only
-	// json = fs.readFileSync('user.json', { encoding: 'utf-8'})
-	// loading.style.display = 'none'
-	// onJson(null, JSON.parse(json).children[10])
-	// onJson(null, JSON.parse(json))
-})
-
-
-function clone(json) {
-	return JSON.parse(JSON.stringify(json))
-}
-
-function clone2(source, target) {
-	if (!target) target = {};
-
-	if (source.name) target.name = source.name;
-	if (source.size) target.size = source.size;
-	if (source.children) {
-		target.children = [];
-		source.children.forEach( node => {
-			target.children.push(clone2(node, {}))
-		})
-	}
-
-	return target;
-}
+module.exports = descendFS
 
 }()
